feat: streamline BDD test suite to focus on core features
- Consolidate BDD tests to three primary features: generate, reports, verify
- Remove redundant and deprecated test files/features
- Maintain comprehensive test coverage with 95 tests (97.9% of target)
- All tests pass: generate (56 tests), reports (25 tests), verify (14 tests)
- Include both BDD scenarios and Python unit tests for complete coverage
- Clean up test structure for better maintainability and clarity
