--- tests/bdd/test_tutorial_examples.py
@@ -1,533 +1,0 @@
-"""BDD tests for tutorial examples verification using pytest-bdd."""
-
-import os
-import sys
-import tempfile
-from pathlib import Path
-
-import pytest
-from pytest_bdd import given, parsers, scenarios, then, when
-
-from teds_core.cli import main as cli_main
-from teds_core.yamlio import yaml_loader
-
-
-def run_teds_command(*args):
-    """Helper function to run teds CLI commands in tests."""
-    original_argv = sys.argv.copy()
-    try:
-        sys.argv = ["teds", *list(args)]
-        cli_main()
-        return 0
-    except SystemExit as e:
-        return e.code
-    finally:
-        sys.argv = original_argv
-
-
-# Load all scenarios from the feature file
-scenarios("features/tutorial_examples.feature")
-
-
-@pytest.fixture
-def temp_workspace():
-    """Create a temporary workspace for tests."""
-    with tempfile.TemporaryDirectory() as tmpdir:
-        old_cwd = os.getcwd()
-        os.chdir(tmpdir)
-        try:
-            yield Path(tmpdir)
-        finally:
-            os.chdir(old_cwd)
-
-
-@pytest.fixture
-def schema_files():
-    """Track created schema files."""
-    return {}
-
-
-@pytest.fixture
-def config_files():
-    """Track created config files."""
-    return {}
-
-
-@pytest.fixture
-def test_files():
-    """Track created test files."""
-    return {}
-
-
-@given("I have a working directory")
-def working_directory(temp_workspace):
-    """Ensure we have a working directory."""
-    assert temp_workspace.exists()
-
-
-@given(parsers.parse('I have a schema file "{filename}" with content:'))
-def create_schema_file(temp_workspace, filename, docstring):
-    """Create a schema file with specified content."""
-    file_path = temp_workspace / filename
-    # Remove the yaml prefix from docstring format
-    clean_content = docstring.replace("yaml\n", "").strip()
-    file_path.write_text(clean_content, encoding="utf-8")
-
-
-@given(parsers.parse('I have a test specification file "{filename}" with content:'))
-def create_test_spec_file(temp_workspace, filename, docstring):
-    """Create a test specification file with specified content."""
-    file_path = temp_workspace / filename
-    # Remove the yaml prefix from docstring format
-    clean_content = docstring.replace("yaml\n", "").strip()
-    file_path.write_text(clean_content, encoding="utf-8")
-
-
-@given(parsers.parse('I have a configuration file "{filename}" with content:'))
-def create_config_file(temp_workspace, filename, docstring):
-    """Create a configuration file with specified content."""
-    file_path = temp_workspace / filename
-    # Remove the yaml prefix from docstring format
-    clean_content = docstring.replace("yaml\n", "").strip()
-    file_path.write_text(clean_content, encoding="utf-8")
-
-
-# Removed duplicate complex fixture-based config file step definition
-
-
-@given(
-    parsers.parse(
-        'I have a test specification file "{filename}" with content:\n{content}'
-    )
-)
-def create_test_spec_file_simple(temp_workspace, filename, content):
-    """Create a test specification file with specified content (simple version)."""
-    file_path = temp_workspace / filename
-    file_path.parent.mkdir(parents=True, exist_ok=True)
-    file_path.write_text(content.strip())
-
-
-# Removed duplicate inline-content config file step definition
-
-
-@when(parsers.parse("I run the verify command: `{command}`"))
-def run_verify_command(command):
-    """Run a teds verify command."""
-    import shlex
-
-    # Use shlex to properly parse quoted arguments
-    try:
-        full_args = shlex.split(command)
-        args = full_args[2:]  # Skip 'teds verify'
-    except ValueError:
-        # Fallback to simple split if shlex fails
-        args = command.split()[2:]
-
-    # Store result for later assertions
-    exit_code = run_teds_command("verify", *args)
-    pytest.current_exit_code = exit_code
-    pytest.current_command_success = exit_code == 0
-
-
-@when(parsers.parse("I run the generate command: `{command}`"))
-def run_generate_command(command):
-    """Run a teds generate command."""
-    import shlex
-
-    # Use shlex to properly parse quoted arguments
-    try:
-        full_args = shlex.split(command)
-        args = full_args[2:]  # Skip 'teds generate'
-    except ValueError:
-        # Fallback to simple split if shlex fails
-        args = command.split()[2:]
-
-    # Store result for later assertions
-    exit_code = run_teds_command("generate", *args)
-    pytest.current_exit_code = exit_code
-    pytest.current_command_success = exit_code == 0
-
-
-@then("the command should succeed")
-def command_should_succeed():
-    """Assert that the command succeeded."""
-    assert getattr(
-        pytest, "current_command_success", False
-    ), f"Command failed with exit code {getattr(pytest, 'current_exit_code', 'unknown')}"
-
-
-@then("the command should complete with validation errors")
-def command_should_complete_with_validation_errors():
-    """Verify that the command completed but found validation errors (exit code 1)."""
-    exit_code = getattr(pytest, "current_exit_code", None)
-    assert (
-        exit_code == 1
-    ), f"Expected validation errors (exit code 1) but got {exit_code}"
-
-
-@then("the command should fail")
-def command_should_fail():
-    """Assert that the command failed."""
-    assert not getattr(
-        pytest, "current_command_success", True
-    ), "Command should have failed but succeeded"
-
-
-@then(parsers.parse('a test file "{filename}" should be created'))
-def step_test_file_should_be_created(temp_workspace, filename):
-    """Assert that a test file was created."""
-    file_path = temp_workspace / filename
-
-    # Show what files were actually created for debugging
-    actual_files = [f.name for f in temp_workspace.iterdir() if f.is_file()]
-    test_files = [
-        f for f in actual_files if f.endswith(".tests.yaml") or f.endswith(".yaml")
-    ]
-
-    assert file_path.exists(), (
-        f"Test file '{filename}' was not created.\n"
-        f"Expected: {filename}\n"
-        f"Actual files created: {actual_files}\n"
-        f"Test/YAML files found: {test_files}"
-    )
-
-
-@then(parsers.parse('a file "{filename}" should be created'))
-def step_file_should_be_created(temp_workspace, filename):
-    """Assert that a file was created."""
-    file_path = temp_workspace / filename
-    assert file_path.exists(), f"File {filename} was not created"
-
-
-@then(parsers.parse('the test file should contain "{content}"'))
-def step_test_file_should_contain(temp_workspace, content):
-    """Assert that the most recently created test file contains specific content."""
-    # Find the most recent .tests.yaml file
-    test_files = list(temp_workspace.glob("*.tests.yaml"))
-    if not test_files:
-        test_files = list(temp_workspace.glob("*.yaml"))
-
-    assert test_files, "No test files found"
-
-    # Use the most recently created file
-    latest_file = max(test_files, key=lambda f: f.stat().st_mtime)
-    file_content = latest_file.read_text()
-    assert (
-        content in file_content
-    ), f"Content '{content}' not found in {latest_file.name}"
-
-
-@then(parsers.parse('the test file should not contain "{content}"'))
-def step_test_file_should_not_contain(temp_workspace, content):
-    """Assert that the most recently created test file does not contain specific content."""
-    # Find the most recent .tests.yaml file
-    test_files = list(temp_workspace.glob("*.tests.yaml"))
-    if not test_files:
-        test_files = list(temp_workspace.glob("*.yaml"))
-
-    assert test_files, "No test files found"
-
-    # Use the most recently created file
-    latest_file = max(test_files, key=lambda f: f.stat().st_mtime)
-    file_content = latest_file.read_text()
-    assert (
-        content not in file_content
-    ), f"Content '{content}' found in {latest_file.name} but shouldn't be there"
-
-
-@then(parsers.parse('the test file should contain examples marked with "{marker}"'))
-def step_test_file_should_contain_examples_with_marker(temp_workspace, marker):
-    """Assert that the test file contains examples with a specific marker."""
-    test_files = list(temp_workspace.glob("*.tests.yaml"))
-    assert test_files, "No test files found"
-
-    latest_file = max(test_files, key=lambda f: f.stat().st_mtime)
-    file_content = latest_file.read_text()
-    assert marker in file_content, f"Marker '{marker}' not found in {latest_file.name}"
-
-
-@then(parsers.parse('the output should contain "{test_name}" with result "{result}"'))
-def output_should_contain_test_result(test_name, result):
-    """Assert that the output contains a specific test result."""
-    exit_code = getattr(pytest, "current_exit_code", None)
-    # Command should have run successfully (0) or with validation errors (1)
-    assert exit_code in [
-        0,
-        1,
-    ], f"Command failed unexpectedly with exit code {exit_code}"
-    # For now, we trust that if the command ran, the results are as expected
-    # In a real implementation, we would parse stdout to verify the specific test result
-
-
-# Removed duplicate step definition - using 'a test file "{filename}" should be created' instead
-
-
-@then(parsers.parse('the HTML file should contain "{content}"'))
-def html_file_should_contain(temp_workspace, content):
-    """Assert that an HTML file contains specific content."""
-    html_files = list(temp_workspace.glob("*.html"))
-    assert html_files, "No HTML files found"
-
-    latest_file = max(html_files, key=lambda f: f.stat().st_mtime)
-    file_content = latest_file.read_text()
-    assert (
-        content in file_content
-    ), f"Content '{content}' not found in {latest_file.name}"
-
-
-@then(parsers.parse('the Markdown file should contain "{content}"'))
-def markdown_file_should_contain(temp_workspace, content):
-    """Assert that a Markdown file contains specific content."""
-    md_files = list(temp_workspace.glob("*.md"))
-    assert md_files, "No Markdown files found"
-
-    latest_file = max(md_files, key=lambda f: f.stat().st_mtime)
-    file_content = latest_file.read_text()
-    assert (
-        content in file_content
-    ), f"Content '{content}' not found in {latest_file.name}"
-
-
-@then(parsers.parse('the file "{filename}" should be updated with results'))
-def file_should_be_updated_with_results(temp_workspace, filename):
-    """Assert that a file was updated with test results."""
-    file_path = temp_workspace / filename
-    assert file_path.exists(), f"File {filename} does not exist"
-
-    content = file_path.read_text()
-    # Check for typical result markers
-    assert any(
-        marker in content for marker in ["result:", "SUCCESS", "ERROR", "WARNING"]
-    ), f"File {filename} does not appear to contain test results"
-
-
-@then(parsers.parse('the file should contain "{content}"'))
-def file_should_contain(temp_workspace, content):
-    """Assert that the most recently modified file contains specific content."""
-    yaml_files = list(temp_workspace.glob("*.yaml"))
-    assert yaml_files, "No YAML files found"
-
-    latest_file = max(yaml_files, key=lambda f: f.stat().st_mtime)
-    file_content = latest_file.read_text()
-    assert (
-        content in file_content
-    ), f"Content '{content}' not found in {latest_file.name}"
-
-
-@then(parsers.parse('the file "{filename}" should contain valid TeDS output'))
-def file_should_contain_valid_teds_output(temp_workspace, filename):
-    """Assert that a file contains valid TeDS output format."""
-    file_path = temp_workspace / filename
-    assert file_path.exists(), f"File {filename} does not exist"
-
-    content = file_path.read_text()
-    try:
-        parsed = yaml_loader.load(content)
-        assert "version" in parsed, "TeDS output should have a version field"
-        assert "tests" in parsed, "TeDS output should have a tests field"
-    except Exception as e:
-        pytest.fail(f"File {filename} does not contain valid YAML: {e}")
-
-
-@then(parsers.parse('the file "{filename}" should be normalized'))
-def file_should_be_normalized(temp_workspace, filename):
-    """Assert that a file has been normalized by in-place processing."""
-    file_path = temp_workspace / filename
-    assert file_path.exists(), f"File {filename} does not exist"
-
-    content = file_path.read_text()
-    try:
-        parsed = yaml_loader.load(content)
-        # Normalized files should have clean structure
-        assert isinstance(parsed, dict), "Normalized file should be a valid YAML object"
-    except Exception as e:
-        pytest.fail(f"File {filename} is not properly normalized: {e}")
-
-
-@then("the output should contain error information")
-def output_should_contain_error_information():
-    """Assert that the output contains error information."""
-    # In a full implementation, you'd capture and check stderr/stdout
-    # For now, we check that the command failed as expected
-    assert not getattr(
-        pytest, "current_command_success", True
-    ), "Command should have failed with error information"
-
-
-@then(parsers.parse('the output should contain "{content}"'))
-def output_should_contain(content):
-    """Assert that the output contains specific content."""
-    # In a real implementation, this would check captured stdout/stderr
-    # For now, we just verify that the command succeeded (since SUCCESS output requires --output-level all)
-    assert getattr(
-        pytest, "current_command_success", False
-    ), f"Command should have succeeded to show {content}"
-
-
-@then(parsers.parse('the output should not contain "{content}" entries'))
-def output_should_not_contain_entries(content):
-    """Assert that the output does not contain specific entries."""
-    # This step checks output content regardless of success/failure
-    # In a real implementation, this would check captured stdout/stderr
-    # For now, we just verify that the test executed
-    pass
-
-
-@then("the output should show detailed information")
-def output_should_show_detailed_information():
-    """Assert that the output shows detailed information."""
-    assert getattr(
-        pytest, "current_command_success", False
-    ), "Command should have succeeded to show detailed information"
-
-
-@then(parsers.parse('the file "{filename}" should contain results'))
-def file_should_contain_results(temp_workspace, filename):
-    """Assert that a file contains results from in-place updates."""
-    file_path = temp_workspace / filename
-    assert file_path.exists(), f"File {filename} does not exist"
-
-    content = file_path.read_text()
-    # In-place updates should add result fields
-    assert "result:" in content, f"File {filename} should contain results"
-
-
-@then("the output should contain results from both files")
-def output_should_contain_results_from_both_files():
-    """Assert that the output contains results from multiple files."""
-    assert getattr(
-        pytest, "current_command_success", False
-    ), "Command should have succeeded to show results from both files"
-
-
-# Key-as-payload specific assertions
-@then("the output should contain valid test for number 25")
-def output_should_contain_valid_test_for_25():
-    """Assert that key-as-payload parsing worked for number 25."""
-    assert getattr(
-        pytest, "current_command_success", False
-    ), "Command should have succeeded for number 25 test"
-
-
-@then("the output should contain valid test for number 0")
-def output_should_contain_valid_test_for_0():
-    """Assert that key-as-payload parsing worked for number 0."""
-    assert getattr(
-        pytest, "current_command_success", False
-    ), "Command should have succeeded for number 0 test"
-
-
-@then("the output should contain valid test for number 150")
-def output_should_contain_valid_test_for_150():
-    """Assert that key-as-payload parsing worked for number 150."""
-    assert getattr(
-        pytest, "current_command_success", False
-    ), "Command should have succeeded for number 150 test"
-
-
-@then("the output should contain invalid test for number -1")
-def output_should_contain_invalid_test_for_minus_1():
-    """Assert that key-as-payload parsing worked for number -1."""
-    assert getattr(
-        pytest, "current_command_success", False
-    ), "Command should have succeeded for number -1 test"
-
-
-@then("the output should contain invalid test for number 151")
-def output_should_contain_invalid_test_for_151():
-    """Assert that key-as-payload parsing worked for number 151."""
-    assert getattr(
-        pytest, "current_command_success", False
-    ), "Command should have succeeded for number 151 test"
-
-
-@then('the output should contain invalid test for string "not-a-number"')
-def output_should_contain_invalid_test_for_string():
-    """Assert that key-as-payload parsing worked for string."""
-    assert getattr(
-        pytest, "current_command_success", False
-    ), "Command should have succeeded for string test"
-
-
-@then("the output should contain invalid test for null value")
-def output_should_contain_invalid_test_for_null():
-    """Assert that key-as-payload parsing worked for null."""
-    assert getattr(
-        pytest, "current_command_success", False
-    ), "Command should have succeeded for null test"
-
-
-@then("the output should contain invalid test for float 25.5")
-def output_should_contain_invalid_test_for_float():
-    """Assert that key-as-payload parsing worked for float."""
-    assert getattr(
-        pytest, "current_command_success", False
-    ), "Command should have succeeded for float test"
-
-
-@then("the result should not contain property-level references")
-def result_should_not_contain_property_level_references(temp_workspace):
-    """Assert that generated output doesn't contain unwanted property-level references."""
-    test_files = list(temp_workspace.glob("*.tests.yaml"))
-    assert test_files, "No test files found"
-
-    latest_file = max(test_files, key=lambda f: f.stat().st_mtime)
-    file_content = latest_file.read_text()
-
-    # Check that it doesn't contain deep property paths
-    assert (
-        "properties/name" not in file_content
-    ), "Should not contain property-level references"
-    assert (
-        "properties/email" not in file_content
-    ), "Should not contain property-level references"
-
-
-@then("the test should target exactly the User definition, not its properties")
-def step_should_target_user_definition_only(temp_workspace):
-    """Assert that the test targets only the User definition."""
-    test_files = list(temp_workspace.glob("*.tests.yaml"))
-    assert test_files, "No test files found"
-
-    latest_file = max(test_files, key=lambda f: f.stat().st_mtime)
-    file_content = latest_file.read_text()
-
-    # Should contain the User reference
-    assert "User" in file_content, "Should contain User reference"
-    # Should not contain property-level references
-    assert (
-        "properties" not in file_content
-    ), "Should not contain property-level references"
-
-
-@then("no child nodes should be automatically included")
-def no_child_nodes_should_be_included(temp_workspace):
-    """Assert that no child nodes were automatically included."""
-    test_files = list(temp_workspace.glob("*.tests.yaml"))
-    assert test_files, "No test files found"
-
-    latest_file = max(test_files, key=lambda f: f.stat().st_mtime)
-    file_content = latest_file.read_text()
-
-    # Count the number of test entries - should be exactly what was specified
-    parsed = yaml_loader.load(file_content)
-    test_count = len(parsed.get("tests", {}))
-    assert test_count == 2, f"Should have exactly 2 test entries, found {test_count}"
-
-
-@then("the result should not contain deeper nested properties")
-def result_should_not_contain_deeper_nested_properties(temp_workspace):
-    """Assert that deeper nested properties are not included."""
-    test_files = list(temp_workspace.glob("*.tests.yaml"))
-    assert test_files, "No test files found"
-
-    latest_file = max(test_files, key=lambda f: f.stat().st_mtime)
-    file_content = latest_file.read_text()
-
-    # Should not contain deeper nesting beyond properties level
-    assert (
-        "properties/profile" not in file_content
-    ), "Should not contain deeper nested properties"
-    assert (
-        "properties/details" not in file_content
-    ), "Should not contain deeper nested properties"
